requirements.txt
# Streamlit and UI
streamlit>=1.28.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Validation
validators>=0.22.0

# Environment and Configuration
python-dotenv>=1.0.0

# AI/ML (for agent integration)
anthropic>=0.18.0

# Data Processing
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
